(WIP) Added search functionality for date range and icons for delete/edit buttons.

+++ frontend/src/app/pins-list-page/pins-list-page.component.ts
@@ -5,6 +5,11 @@ import {MatTableDataSource} from "@angular/material/table";
 import {Poi} from "../models/poi.model";
 import {MatDialog} from "@angular/material/dialog";
 import {PinsEditDialogComponent} from "./pins-list-form/pins-edit-dialog.component";
+import {MatSort} from "@angular/material/sort";
+
+
+import {BreakpointObserver, Breakpoints} from '@angular/cdk/layout';
+
 
 @Component({
     selector: 'app-pins-list-page',
@@ -14,19 +19,29 @@ import {PinsEditDialogComponent} from "./pins-list-form/pins-edit-dialog.compone
 })
 
 export class PinsListPageComponent implements OnInit {
-
     pois: Poi[] = [];
 
     dataSource: MatTableDataSource<Poi>;
-    displayedColumns: string[] = ['name', 'description', 'lat', 'lng'];
+    displayedColumns: string[] = ['name', 'description', 'category', 'lat', 'lng', 'created'];
+
 
     @ViewChild(MatPaginator) paginator!: MatPaginator;
+    @ViewChild(MatSort, {static: false}) sort!: MatSort;
 
-    constructor(private poiService: PoiService, private dialog: MatDialog) {
+    isHandset: boolean = false;
+
+    constructor(private poiService: PoiService, private dialog: MatDialog, private breakpointObserver: BreakpointObserver) {
         this.dataSource = new MatTableDataSource(this.pois);
     }
 
     ngOnInit(): void {
+        this.breakpointObserver.observe([
+            Breakpoints.HandsetPortrait,
+            Breakpoints.HandsetLandscape,
+        ]).subscribe(result => {
+            this.isHandset = result.matches;
+            console.log('Is handset:', this.isHandset); // Ska logga true eller false när storleken på skärmen ändras
+        });
         this.loadPois();
     }
 
@@ -35,10 +50,29 @@ export class PinsListPageComponent implements OnInit {
             this.pois = data;
             this.dataSource = new MatTableDataSource(this.pois);
             this.dataSource.paginator = this.paginator;
+            this.dataSource.sort = this.sort;
+
+            this.dataSource.sortingDataAccessor = (poi: Poi, sortHeaderId: string) => {
+                if (sortHeaderId in poi) {
+                    switch (sortHeaderId) {
+                        case 'created':
+                            // Kontrollera först om 'created' är en sträng för att undvika runtime-fel
+                            // om 'created' inte är en sträng, borde du hantera det på ett lämpligt sätt
+                            return new Date((poi as any)[sortHeaderId]);
+                        default:
+                            return (poi as any)[sortHeaderId];
+                    }
+                }
+                throw new Error(`Property '${sortHeaderId}' is not a valid key of Poi.`);
+            };
         });
     }
 
-    deletePoi(id: number) {
+    deletePoi(id: number | undefined) {
+        if (id === undefined) {
+            console.error('Försöker radera en POI utan ett ID');
+            return;
+        }
         this.poiService.deletePoi(id).subscribe(
             data => {
                 console.log('deleted response', data);
@@ -47,15 +81,16 @@ export class PinsListPageComponent implements OnInit {
         )
     }
 
-
     editPoi(poi: Poi) {
+        const originalCreatedDate = poi.created; // Spara det ursprungliga createdDate
+
         const dialogRef = this.dialog.open(PinsEditDialogComponent, {
-            data: {poi},
+            data: {poi, originalCreatedDate},
         });
 
         dialogRef.afterClosed().subscribe((result: Poi) => {
             if (result && result.id !== undefined) {
-                this.updatePoi(result.id, result);
+                this.updatePoi(result.id, {...result, created: originalCreatedDate});
             }
         });
     }
@@ -70,12 +105,6 @@ export class PinsListPageComponent implements OnInit {
                 console.error('Error updating POI', error);
             }
         });
-    }
-
-
-    applyFilter(event: any) {
-        const filterValue = (event.target as HTMLInputElement).value;
-        this.dataSource.filter = filterValue.trim().toLowerCase();
     }
 
 }
